🗜️ refactor: change file structure
