--- student_app/app/src/main/java/io/github/kabirnayeem99/dumarketingstudent/presentation/base/BaseActivity.kt
@@ -1,29 +1,0 @@
-package io.github.kabirnayeem99.dumarketingstudent.presentation.base
-
-import android.os.Bundle
-import android.view.inputmethod.InputMethodManager
-import androidx.annotation.LayoutRes
-import androidx.appcompat.app.AppCompatActivity
-import androidx.databinding.DataBindingUtil
-import androidx.databinding.ViewDataBinding
-
-
-abstract class BaseActivity<B : ViewDataBinding> : AppCompatActivity() {
-    protected lateinit var binding: B
-
-    @get:LayoutRes
-    protected abstract val layout: Int
-
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = DataBindingUtil.setContentView(this, layout)
-        binding.lifecycleOwner = this
-    }
-
-
-    open fun hideKeyboardFrom() {
-        val imm: InputMethodManager = getSystemService(INPUT_METHOD_SERVICE) as InputMethodManager
-        imm.hideSoftInputFromWindow(binding.root.windowToken, 0)
-    }
-}
