--- student_app/app/src/main/java/io/github/kabirnayeem99/dumarketingstudent/presentation/base/BaseFragment.kt
@@ -1,32 +1,0 @@
-package io.github.kabirnayeem99.dumarketingstudent.presentation.base
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.annotation.LayoutRes
-import androidx.databinding.DataBindingUtil
-import androidx.databinding.ViewDataBinding
-import androidx.fragment.app.Fragment
-
-
-abstract class BaseFragment<B : ViewDataBinding> : Fragment() {
-
-    protected lateinit var binding: B
-
-
-    @get:LayoutRes
-    protected abstract val layout: Int
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        binding = DataBindingUtil.inflate<B>(inflater, layout, container, false).also {
-            it.lifecycleOwner = this.viewLifecycleOwner
-        }
-        return binding.root
-    }
-
-}
