--- student_app/app/src/main/java/io/github/kabirnayeem99/dumarketingstudent/presentation/base/ViewModelFactory.kt
@@ -1,24 +1,0 @@
-package io.github.kabirnayeem99.dumarketingstudent.presentation.base
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import javax.inject.Inject
-import javax.inject.Provider
-
-class ViewModelFactory @Inject constructor(
-    private val viewModelsMap: Map<Class<out ViewModel>, @JvmSuppressWildcards Provider<ViewModel>>
-) : ViewModelProvider.Factory {
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        val creator = viewModelsMap[modelClass]
-            ?: viewModelsMap.entries.firstOrNull {
-                modelClass.isAssignableFrom(it.key)
-            }?.value ?: throw IllegalArgumentException("Unknown ViewModel class $modelClass")
-
-        return try {
-            @Suppress("UNCHECKED_CAST")
-            creator.get() as T
-        } catch (e: Exception) {
-            throw RuntimeException(e)
-        }
-    }
-}
