+++ student_app/app/src/main/java/io/github/kabirnayeem99/dumarketingstudent/presentation/faculty/FacultyViewModel.kt
@@ -1,4 +1,4 @@
-package io.github.kabirnayeem99.dumarketingstudent.presentation.viewmodel
+package io.github.kabirnayeem99.dumarketingstudent.presentation.faculty
 
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
