+++ student_app/app/src/main/java/io/github/kabirnayeem99/dumarketingstudent/presentation/gallery/GalleryDataAdapter.kt
@@ -1,4 +1,4 @@
-package io.github.kabirnayeem99.dumarketingstudent.presentation.adapters
+package io.github.kabirnayeem99.dumarketingstudent.presentation.gallery
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
