+++ student_app/app/src/main/java/io/github/kabirnayeem99/dumarketingstudent/presentation/gallery/GalleryFragment.kt
@@ -1,7 +1,6 @@
-package io.github.kabirnayeem99.dumarketingstudent.presentation.fragments
+package io.github.kabirnayeem99.dumarketingstudent.presentation.gallery
 
 import android.os.Bundle
-import android.view.View
 import android.view.animation.ScaleAnimation
 import androidx.fragment.app.activityViewModels
 import androidx.lifecycle.observe
@@ -9,12 +8,10 @@ import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.GridLayoutManager
 import dagger.hilt.android.AndroidEntryPoint
 import io.github.kabirnayeem99.dumarketingstudent.R
-import io.github.kabirnayeem99.dumarketingstudent.databinding.FragmentGalleryBinding
-import io.github.kabirnayeem99.dumarketingstudent.presentation.adapters.GalleryDataAdapter
-import io.github.kabirnayeem99.dumarketingstudent.presentation.base.BaseFragment
+import io.github.kabirnayeem99.dumarketingstudent.common.base.BaseFragment
 import io.github.kabirnayeem99.dumarketingstudent.common.util.Resource
 import io.github.kabirnayeem99.dumarketingstudent.common.util.showSnackBar
-import io.github.kabirnayeem99.dumarketingstudent.presentation.viewmodel.GalleryViewModel
+import io.github.kabirnayeem99.dumarketingstudent.databinding.FragmentGalleryBinding
 import me.everything.android.ui.overscroll.OverScrollDecoratorHelper
 import timber.log.Timber
 import javax.inject.Inject
@@ -22,7 +19,7 @@ import javax.inject.Inject
 @AndroidEntryPoint
 class GalleryFragment : BaseFragment<FragmentGalleryBinding>() {
 
-    override val layout: Int
+    override val layoutRes: Int
         get() = R.layout.fragment_gallery
 
     private val galleryViewModel: GalleryViewModel by activityViewModels()
@@ -33,13 +30,12 @@ class GalleryFragment : BaseFragment<FragmentGalleryBinding>() {
     private val galleryAdapter: GalleryDataAdapter by lazy {
         GalleryDataAdapter {
             galleryViewModel.setSelectedImageUrl(it.imageUrl)
-         findNavController()
+            findNavController()
                 .navigate(R.id.action_galleryFragment_to_imageDetailsFragment)
         }
     }
 
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
+    override fun onCreated(savedInstanceState: Bundle?) {
         setUpRecyclerView()
     }
