+++ student_app/app/src/main/java/io/github/kabirnayeem99/dumarketingstudent/presentation/gallery/ImageDetailsFragment.kt
@@ -1,16 +1,14 @@
-package io.github.kabirnayeem99.dumarketingstudent.presentation.fragments
+package io.github.kabirnayeem99.dumarketingstudent.presentation.gallery
 
 import android.os.Bundle
-import android.view.View
 import android.view.animation.ScaleAnimation
 import androidx.fragment.app.activityViewModels
 import com.bumptech.glide.Glide
 import dagger.hilt.android.AndroidEntryPoint
 import io.github.kabirnayeem99.dumarketingstudent.R
-import io.github.kabirnayeem99.dumarketingstudent.databinding.FragmentImageDetailsBinding
-import io.github.kabirnayeem99.dumarketingstudent.presentation.base.BaseFragment
+import io.github.kabirnayeem99.dumarketingstudent.common.base.BaseFragment
 import io.github.kabirnayeem99.dumarketingstudent.common.util.showSnackBar
-import io.github.kabirnayeem99.dumarketingstudent.presentation.viewmodel.GalleryViewModel
+import io.github.kabirnayeem99.dumarketingstudent.databinding.FragmentImageDetailsBinding
 import timber.log.Timber
 import javax.inject.Inject
 
@@ -23,12 +21,11 @@ class ImageDetailsFragment : BaseFragment<FragmentImageDetailsBinding>() {
 
     private val galleryViewModel: GalleryViewModel by activityViewModels()
 
-    override val layout: Int
+    override val layoutRes: Int
         get() = R.layout.fragment_image_details
 
 
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
+    override fun onCreated(savedInstanceState: Bundle?) {
         galleryViewModel.selectedImageUrl.observe(viewLifecycleOwner, { imageUrl ->
             loadUi(imageUrl)
         })
