+++ student_app/app/src/main/java/io/github/kabirnayeem99/dumarketingstudent/presentation/routine/RoutineDataAdapter.kt
@@ -1,4 +1,4 @@
-package io.github.kabirnayeem99.dumarketingstudent.presentation.adapters
+package io.github.kabirnayeem99.dumarketingstudent.presentation.routine
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
