+++ student_app/app/src/main/res/layout/activity_main.xml
@@ -8,7 +8,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:animateLayoutChanges="true"
-        tools:context=".presentation.activities.MainActivity">
+        tools:context=".presentation.MainActivity">
 
 
         <androidx.constraintlayout.widget.ConstraintLayout
