+++ student_app/app/src/main/res/layout/fragment_image_details.xml
@@ -7,7 +7,7 @@
     <RelativeLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.fragments.ImageDetailsFragment">
+        tools:context=".presentation.gallery.ImageDetailsFragment">
 
         <androidx.constraintlayout.widget.ConstraintLayout
             android:layout_width="match_parent"
