+++ student_app/app/src/main/res/values/strings.xml
@@ -7,4 +7,6 @@
 \n With a view to exposing the students to the out-sidethe- campus world, the department organizes guest lectures, seminars etc. inviting reputed personalities from the public and private sectors. The department occasionally organizes study tour within Bangladesh and abroad for the students of final year. \n Journal of Marketing is a half-yearly referred journal of the Department of Marketing which publishes research-based articles contributed by the academicians, educationists and professionals. \n</string>
     <string name="dummy_button">Dummy Button</string>
     <string name="dummy_content">DUMMY\nCONTENT</string>
+    <string name="today_s_classes">Today\'s Classes</string>
+    <string name="recent_notices">Recent Notices</string>
 </resources>
